feat(core): add comlink controller and channel store (#141)
### Description

(note: this is the first of 4 stacked PRs. To see the end result, go to [the demo](https://sdk-kitchensink-react-m8t6nfdel.sanity.dev/comlink-demo) ☺️ )

This PR adds a store and actions for a comlink controller and channels, which are intended to be used by apps that load an iframe and need to communicate with them. You can think of these as being used by the parent. They're ultimately used in a `useFrameConnection` hook, added in [this PR](https://github.com/sanity-io/sdk/pull/153/files). 

### What to review

- The `getOrCreate` type functions are pretty optimistic, and meant to be used with as little fuss as possible. Do we see any issues with this approach of creating channels / re-creating controllers without much fanfare?
- If we think about how these are used, channels are likely to be long-lived (a user hanging out in an app, switching back and forth, and so on). For this reason, we count usage and release the channel, but we do not destroy it (channels will be destroyed in the `dispose` method of the store). Are we okay with this?

### Testing

100% test coverage but let me know if there's cases we may be missing.

### Fun gif

![parent](https://media2.giphy.com/media/v1.Y2lkPTc5MGI3NjExeG11OXhiM2s4d2RrbWp5a2c1eDF5bGNhYmNtbWRrbDdqNmc2eTZjayZlcD12MV9pbnRlcm5hbF9naWZfYnlfaWQmY3Q9Zw/UIRlVJe5ksC6Q/giphy.webp)