feat(core): add comlink node store (#156)
### Description

(note: this is the second of 4 stacked PRs. To see the end result, go to [the demo](https://sdk-kitchensink-react-m8t6nfdel.sanity.dev/comlink-demo) 😌  )

This PR adds a store and actions for Comlink nodes, which are intended to be used by apps loaded within an iframe. You can think of these as being used by the "child" of a Comlink connection. They're ultimately used in a `useWindowConnection` hook, added in [this PR](https://github.com/sanity-io/sdk/pull/153/files). 

### What to review
- This implementation is similar to how we think of [channels in the "parent" store](https://github.com/sanity-io/sdk/pull/141).
- The `getOrCreate` type functions are pretty optimistic, and meant to be used with as little fuss as possible. Do we see any issues with this approach of creating nodes without much fanfare?
- If we think about how these are used, nodes are likely to be long-lived (a user hanging out in an app, switching back and forth, and so on). For this reason, we count usage and release the node, but we do not destroy it (nodes will be stopped in the `dispose` method of the store). Are we okay with this?

### Testing

100% test coverage but let me know if there's cases we may be missing.

### Fun gif

![child](https://media3.giphy.com/media/v1.Y2lkPTc5MGI3NjExODk5ZHpwNXpuMzNmOWU1NWtjdmFvYWt5MGJ1aXdjdGY0amlzdXZuaSZlcD12MV9pbnRlcm5hbF9naWZfYnlfaWQmY3Q9Zw/sR91D133W02D6/200.webp)