fix(core): align release order with studio release order (#534)
### Description

This PR updates the release perspective handling in @sdk/core to match the behavior in the Studio -- it seems like better DX that these behaviors are the same across applications. Additionally, it exports the `ReleasePerspective` type and filters out archived and published releases from the active releases list.

### What to review

- The changes to `getPerspectiveState.ts` and `releasesStore.ts` to ensure the sorting and filtering logic is correct
- The updated tests in `getPerspectiveState.test.ts` and `releasesStore.test.ts` to verify the new behavior
- The export of the `ReleasePerspective` type in the core exports

### Testing

Updated existing tests to account for the reversed order of releases. The tests now correctly expect the releases to be in reverse order to match the UI behavior elsewhere.

### Fun gif
![](https://media4.giphy.com/media/v1.Y2lkPTc5MGI3NjExejMwbW1sankzc3Zuc3BiMzV6aWZ0cTgwdWNwejg3bDZtaXc2ZGo3bSZlcD12MV9pbnRlcm5hbF9naWZfYnlfaWQmY3Q9Zw/13hzdQ3QCID172/giphy.gif)