feat(kitchensink): add additional readonly hooks to releases route (#535)
### Description

This PR cleans up the Releases route in Kitchensink a bit to make it a bit more readable and allow for easier debugging. I also added a change to the DocumentPreview component an onClick handler to support document selection (not currently in play 😅 )

If you test this manually, note that the useDocument, usePreview, and useProjection hooks are not currently working. That'll be a later PR!

### What to review
Are we happy? Anything strange in Kitchensink?

### Testing
No tests for Kitchensink... yet 😏 

### Fun gif
![](https://media0.giphy.com/media/v1.Y2lkPTc5MGI3NjExbDFmcHIzM2lzYW5yc2diYTAwb2Q4ZmFsaWYwazkxYmppam9pa3RseSZlcD12MV9pbnRlcm5hbF9naWZfYnlfaWQmY3Q9Zw/Rk8wCrJCrjRJ2MyLrb/giphy.gif)