fix(core): correctly handle state changes during onSubscribe in createStateSourceAction
This fixes a quite subtle bug: Any state changes happening in `onSubscribe`
would be not be part of the source.observable. The reason for this is
that the implementation would do:

1. First emit the current value by looking at state.get()
2. Invoke the onSubscribe hook.
3. Listen to the observable of state, but skip the first value
   because we expect it to emit the current state first.

However, this means that if (2) actually _changes_ the state synchronously
then that's the value that is skipped in (3) happens to be the new state.

This changes the implementation to be based around an observable
which contains the actual values. This is both shorter in code
and also happens to avoid calling `getCurrent` multiple times.
Previously `getCurrent` would be invoked twice: Once for detecting
that a value had changed and once again when producing the value.