fix(react): ensure useDocumentProjection reuses state source for all strings (#669)
### Description
`useDocumentProjection` was actually a bit finicky about the `projection` being passed to it. We had a ticket about it looping forever.

My best guess about what was happening was this:
1. Any helper functions, or even template literals, created new references passed to `useDocumentProjection`
2. `useDocumentProjection` would then call `getProjectionState`. It always returns a new `StateSource`, which returns a new `subscribe` function
3. `useSyncExternalStore` receives the new `subscribe` function, triggering a re-render
4. re-render causes the loop to start all over again

This PR adds some memoization to maintain consistency in projection strings and state sources.

We didn't catch this in casual clicking around because `defineProjection` was doing some of the memoization work for us, I think. To address this, I added some common ways people would use `useDocumentProjection` and also some e2e tests against it.

### What to review
Any risks of over memoization or staleness?

### Testing
e2e tests (I think this is hard to capture in React tests and we still have 100% coverage).

To see the OLD behavior, 
1. `git fetch`
2. `git checkout main`
3. `git checkout origin/fix/sdk-660/ensure-doc-projections-reuse-state-source apps/kitchensink-react/src/DocumentCollection/DocumentProjectionRoute.tsx`
4. `pnpm dev`
5. go to https://www.sanity.io/@oblZgbTFj?dev=http%3A%2F%2Flocalhost%3A3333
6. go to https://www.sanity.io/@oblZgbTFj/application/__dev/document-projection
7. Click around some of the many buttons, you'll see things hanging.

This PR fixes that.