feat: support using the same instance for different datasets (#632)
This introduces the concept of a "key" for a store. The main goal here
is to enable the support for us using a single SanityInstance across all
the stores, but still have separate stores per projectId/dataset. This
is accomplished by saying that each store is determined by a "key" which
can be dynamically built from the options. We first invoke the key logic
and then dispatches to the right store. For now have projectId/dataset
keys, but in the future we expect this to be per-resource.

This also means that it's possible to use a single SanityInstance for
working with multiple projectId/dataset. If you pass in an explicit
projectId/dataset then it will no longer care about the configuration on
the SanityInstance.